[nemo-qml-plugin-email] Improve drafts handling
Some services(e.g gmail, outlook) automatically create emails in the
send folder when using their smtp and if after sending the imap client
does not append a email to the same folder.
When sending a draft that exists in the server side,
if we copy the existent email to the send folder it will cause a duplicate
due to the automatic creation by the server, with this commit we always create
new emails from existent drafts and then delete the drafts after enqueing them for
sending.