[nemo-qml-plugin-email] Introduce hasPersistentConnection property and model role.
hasPersistentConnection property indicates that one or more accounts has a
persistent connection to the server(always online), this can be used if
applications want to show a global status of last changes to accounts.
hasPersistentConnection model role returns true if the account is in
always up to date mode, false otherwise.